fix: test code 분리 및 수정
