+++ apps/core-api/test/auth.e2e-spec.ts
